Optional positional arguments now working
