+++ Richiban.Cmdr.sln
@@ -17,8 +17,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Richiban.Cmdr.Tests", "Rich
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Richiban.Cmdr.Samples", "Richiban.Cmdr.Samples\Richiban.Cmdr.Samples.csproj", "{ED6D84AA-4121-4202-95F4-1F00882EF264}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Richiban.Cmdr.TestBench", "Richiban.Cmdr.TestBench\Richiban.Cmdr.TestBench.csproj", "{B5A3B2F6-6CB6-41D8-80E9-9E6E4101CCBB}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -37,10 +35,6 @@ Global
 		{ED6D84AA-4121-4202-95F4-1F00882EF264}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ED6D84AA-4121-4202-95F4-1F00882EF264}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ED6D84AA-4121-4202-95F4-1F00882EF264}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B5A3B2F6-6CB6-41D8-80E9-9E6E4101CCBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B5A3B2F6-6CB6-41D8-80E9-9E6E4101CCBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B5A3B2F6-6CB6-41D8-80E9-9E6E4101CCBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B5A3B2F6-6CB6-41D8-80E9-9E6E4101CCBB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
